x86/efi: Map EFI_BOOT_SERVICES_* regions as unaccessible
In order to catch illegal access to boot services regions we want all
accesses to these regions to page fault and hence we map these regions
as unaccessible. We also use _PAGE_SPECIAL flag to mark these mappings
as EFI_BOOT_SERVICES_* regions. We check for this bit in the page fault
handler because we will not fix any other page faults except "illegal
access to boot services regions". Although page flag bits are
constrained, using a page flag bit for this puspose should be OK because
we use this bit only in efi_pgd and not in any other pgd (specially
swapper_pgd).

Signed-off-by: Sai Praneeth Prakhya <sai.praneeth.prakhya@intel.com>
Cc: Lee, Chun-Yi <jlee@suse.com>
Cc: Borislav Petkov <bp@alien8.de>
Cc: Fenghua Yu <fenghua.yu@intel.com>
Cc: Ravi Shankar <ravi.v.shankar@intel.com>
Cc: Matt Fleming <matt@codeblueprint.co.uk>
Cc: Ard Biesheuvel <ard.biesheuvel@linaro.org>
CC: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>