x86/efi: remove __init attribute from memory mapping functions
All these functions will be used by the page fault handler to fixup
access to BOOT_SERVICES_* regions. The page fault handler doesn't have
an __init attribute, but these functions have it and thus warns at build
(WARNING: modpost: Found 4 section mismatch(es).) and to fix that we
remove __init attribute for all these functions.

In order to not keep these functions needlessly when
"CONFIG_EFI_BOOT_SERVICES_WARN" is not selected, we add a new
__init_fixup attribute whose value changes based on whether the config
option is selected or not.

Signed-off-by: Sai Praneeth Prakhya <sai.praneeth.prakhya@intel.com>
Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Cc: Lee, Chun-Yi <jlee@suse.com>
Cc: Borislav Petkov <bp@alien8.de>
Cc: Matt Fleming <matt@codeblueprint.co.uk>
Cc: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Cc: Ravi Shankar <ravi.v.shankar@intel.com>
Cc: Fenghua Yu <fenghua.yu@intel.com>